feat: Remove custom layout configurations
For now, we remove custom layouts involving ini files and the like. The feature may be brought back, though that will come down to need for the feature, which may very well not present itself.

Also, we remove fetching an updated controller mapping database. There's no need for this app to search for an updated database everytime the controller display is launched.